requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
requests
beautifulsoup4
yfinance
newspaper3k
newsapi-python
spaCy
transformers
torch
pandas
matplotlib
streamlit
plotly
feedparser
